fix: proper routes and structure of home module
