+++ src/app.routes.tsx
@@ -2,20 +2,20 @@ import React from 'react';
 import { Navigate, Route, Routes } from 'react-router-dom';
 
 import AccessoryRoutes from 'app/accessory/routes';
+import AuthRoutes from 'app/auth/routes';
 import BeerRoutes from 'app/beer/routes';
 import CartRoutes from 'app/cart/routes';
-import HomeRoutes from 'app/home-pages/routes';
 import OrderRoutes from 'app/order/routes';
 import SnackRoutes from 'app/snack/routes';
 
-const AuthRoutes = React.lazy(() => import('app/auth/routes'));
+const HomePage = React.lazy(() => import('./app/home/home.page'));
 
 const AppRoutes = () => {
   return (
     <Routes>
       <Route path="/auth/*" element={<AuthRoutes />} />
 
-      <Route path="/home/*" element={<HomeRoutes />} />
+      <Route path="/" element={<HomePage />} />
       <Route path="/beer/*" element={<BeerRoutes />} />
       <Route path="/snacks/*" element={<SnackRoutes />} />
       <Route path="/accessories/*" element={<AccessoryRoutes />} />
